mobile base change, bias between wheel to wheel changed

+++ xbot_core/xbot_driver/include/xbot_driver/packets/core_sensors.hpp
@@ -68,7 +68,7 @@ public:
      short yaw;
      short pitch;
      short roll;
-     unsigned int timestamp;
+     unsigned short timestamp;
   } data;
 
   struct Flags {
